+++ HISTORY.rst
@@ -2,11 +2,12 @@
 History
 =======
 
-0.8.1
------
+unreleased
+----------
 
 Fixed
 ^^^^^
+- `pyfar.dsp.time_crop` did not work correctly for higher channel dimensional inputs (PR #944)
 - A bug where a numpy array being passed as a `pyfar.audio.classes.Signal.sampling_rate` raised `TypeError`.